+++ source/examples/api/omni.isaac.franka/franka_follow_target.py
@@ -10,34 +10,28 @@ from omni.isaac.kit import SimulationApp
 
 simulation_app = SimulationApp({"headless": False})
 
-from omni.isaac.franka.tasks import TargetFollower
-from omni.isaac.franka.controllers import RMPFlowIKSolver
+from omni.isaac.franka.tasks import FollowTarget
+from omni.isaac.franka.controllers import RMPFlowController
 from omni.isaac.core import World
-from omni.isaac.core.utils.extensions import get_extension_id, get_extension_path
+from omni.isaac.franka.controllers import InverseKinematicsSolver
+from omni.isaac.core.utils.rotations import euler_angles_to_quat
+import numpy as np
 
 my_world = World()
-extension_id = get_extension_id("omni.isaac.motion_generation")
-mg_extension_path = get_extension_path(ext_id=extension_id)
-my_task = TargetFollower()
+my_task = FollowTarget()
 my_world.load_task(my_task)
 my_world.reset()
 my_franka = my_world.scene.get_object("my_franka")
-my_controller = RMPFlowIKSolver(
-    name="target_follower_controller",
-    dc_interface=my_world.dc_interface,
-    stage=my_world.stage,
-    robot_prim=my_franka.prim,
-    mg_extension_path=mg_extension_path,
-)
+# my_controller = InverseKinematicsSolver(
+#     name="target_follower_controller",
+#     robot_prim_path=my_franka.prim_path)
+my_controller = RMPFlowController(name="target_follower_controller", robot_prim_path=my_franka.prim_path)
 articulation_controller = my_franka.get_articulation_controller()
 
 i = 0
 while True:
     observations = my_world.get_observations()
-    actions = my_controller.forward(
-        target_end_effector_position=observations["target_cube"]["position"],
-        current_joint_positions=observations["my_franka"]["joint_positions"],
-    )
+    actions = my_controller.forward(target_end_effector_position=observations["target"]["position"])
     articulation_controller.apply_action(actions)
     my_world.step(render=True)
     if i % 2000 == 0: