+++ source/examples/api/omni.isaac.franka/stacking.py
@@ -10,26 +10,23 @@ from omni.isaac.kit import SimulationApp
 
 simulation_app = SimulationApp({"headless": False})
 
-from omni.isaac.franka.tasks import Tower
-from omni.isaac.franka.controllers import RMPFlowTower
+from omni.isaac.franka.tasks import Stacking
+from omni.isaac.franka.controllers import StackingController
 from omni.isaac.core import World
-from omni.isaac.core.utils.extensions import get_extension_id, get_extension_path
 
 my_world = World()
-extension_id = get_extension_id("omni.isaac.motion_generation")
-mg_extension_path = get_extension_path(ext_id=extension_id)
-my_task = Tower()
+my_task = Stacking()
 my_world.load_task(my_task)
 my_world.reset()
-my_franka = my_world.scene.get_object("my_franka")
-my_controller = RMPFlowTower(
+robot_name = my_task.get_robot_name()
+my_franka = my_world.scene.get_object(robot_name)
+my_controller = StackingController(
     name="pick_place_controller",
-    dc_interface=my_world.dc_interface,
-    stage=my_world.stage,
-    robot_prim=my_franka.prim,
-    mg_extension_path=mg_extension_path,
+    gripper_dof_indices=my_franka.gripper.dof_indices,
+    robot_prim_path=my_franka.prim_path,
+    picking_order_cube_names=my_task.get_cube_names(),
+    robot_observation_name=robot_name,
 )
-my_controller.configure(cubes_order=["cube_1", "cube_2"], robot_observation_name="my_franka")
 articulation_controller = my_franka.get_articulation_controller()
 
 i = 0