+++ source/extensions/omni.isaac.franka/omni/isaac/franka/tasks/follow_target.py
@@ -10,9 +10,18 @@ import omni.isaac.core.tasks as tasks
 from omni.isaac.franka import Franka
 
 
-class PickPlace(tasks.PickPlace):
-    def __init__(self) -> None:
+class FollowTarget(tasks.FollowTarget):
+    def __init__(
+        self, target_prim_path="/World/TargetCube", target_name="target", target_position=None, target_orientation=None
+    ) -> None:
         """[summary]
         """
-        tasks.PickPlace.__init__(self, robot=Franka(prim_path="/World/Franka", name="my_franka"))
+        tasks.FollowTarget.__init__(
+            self,
+            robot=Franka(prim_path="/World/Franka", name="my_franka"),
+            target_prim_path=target_prim_path,
+            target_name=target_name,
+            target_position=target_position,
+            target_orientation=target_orientation,
+        )
         return