+++ source/extensions/omni.isaac.motion_generation/omni/isaac/motion_generation/pick_place_controller.py
@@ -6,27 +6,33 @@
 # distribution of this software and related documentation without an express
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
-from omni.isaac.franka.controllers.rmpflow_ik import RMPFlowIKSolver
+from omni.isaac.core.controllers import BaseController
 from omni.isaac.core.utils.types import ArticulationAction
 from omni.isaac.core.utils.rotations import euler_angles_to_quat
 import numpy as np
 
 
-class RMPFlowPickPlace(RMPFlowIKSolver):
+class PickPlaceController(BaseController):
     # TODO: this will need further discussion with buck and SRL before cleaning it up
-    def __init__(self, name, mg_extension_path, dc_interface, stage, robot_prim):
-        super().__init__(
-            name=name,
-            mg_extension_path=mg_extension_path,
-            dc_interface=dc_interface,
-            stage=stage,
-            robot_prim=robot_prim,
-        )
+    def __init__(self, name, ik_solver, gripper_controller, start_picking_height=0.3, event_velocities=None):
+        BaseController.__init__(self, name=name)
         self._event = 0
         self._t = 0
-        self._h1 = 0.3
+        self._h1 = start_picking_height
         self._h0 = None
-        self._event_velocities = [0.08, 0.005, 0.3, 0.0025, 0.001, 0.0025, 1, 0.008, 0.08]
+        self._event_velocities = event_velocities
+        if self._event_velocities is None:
+            self._event_velocities = [0.008, 0.005, 0.3, 0.0025, 0.001, 0.0025, 1, 0.008, 0.08]
+        else:
+            if not isinstance(self._event_velocities, np.ndarray) or not isinstance(self._event_velocities, list):
+                raise Exception("event velocities need to be list or numpy array")
+            elif isinstance(self._event_velocities, np.ndarray):
+                self._event_velocities = self._event_velocities.tolist()
+            if len(self._event_velocities) == 9:
+                raise Exception("event velocities need have length of 9")
+        self._ik_solver = ik_solver
+        self._gripper_controller = gripper_controller
+        self._pause = False
         """
         - Phase 0: Move end_effector above the cube center.
         - Phase 1: Lower end_effector down to encircle the target cube
@@ -36,39 +42,57 @@ class RMPFlowPickPlace(RMPFlowIKSolver):
         - Phase 5: Move end_effector vertically toward goal height.
         - Phase 6: loosen the grip.
         - Phase 7: Move end_effector vertically up again
-        - Phase 8: Move end_effector towards the old xy position. 
+        - Phase 8: Move end_effector towards the old xy position.
         """
         return
 
-    def forward(self, cube_position, cube_orientation, cube_target_position, current_joint_positions):
-        if self._event >= len(self._event_velocities):
+    def is_paused(self):
+        return self._pause
+
+    def get_current_event(self):
+        return self._event
+
+    def forward(
+        self,
+        cube_position,
+        cube_orientation,
+        cube_target_position,
+        current_joint_positions,
+        end_effector_translation_offset=None,
+    ):
+        if end_effector_translation_offset is None:
+            end_effector_translation_offset = np.array([0, 0, 0])
+        if self._pause or self._event >= len(self._event_velocities):
             target_joint_positions = [None] * current_joint_positions.shape[0]
             return ArticulationAction(joint_positions=target_joint_positions)
 
         if self._event < 2:
             self._current_target_x = cube_position[0]
             self._current_target_y = cube_position[1]
-            self._h0 = cube_position[2] - 0.005
+            self._h0 = cube_position[2]
         # TODO: take into account cube orientation
 
         interpolated_xy = self._get_interpolated_xy(
             cube_target_position[0], cube_target_position[1], self._current_target_x, self._current_target_y
         )
         target_height = self._get_target_hs(cube_target_position[2])
-        position_target = np.array([interpolated_xy[0], interpolated_xy[1], target_height])
+        position_target = np.array(
+            [
+                interpolated_xy[0] + end_effector_translation_offset[0],
+                interpolated_xy[1] + end_effector_translation_offset[1],
+                target_height + end_effector_translation_offset[2],
+            ]
+        )
         if self._event == 2:
-            target_joint_positions = [None] * current_joint_positions.shape[0]
-            target_joint_positions[7] = current_joint_positions[7] - 0.005
-            target_joint_positions[8] = current_joint_positions[8] - 0.005
-            target_joint_positions = ArticulationAction(joint_positions=target_joint_positions)
+            target_joint_positions = self._gripper_controller.forward(
+                action="close", current_joint_positions=current_joint_positions
+            )
         elif self._event == 6:
-            target_joint_positions = [None] * current_joint_positions.shape[0]
-            target_joint_positions[7] = current_joint_positions[7] + 0.005
-            target_joint_positions[8] = current_joint_positions[8] + 0.005
-            target_joint_positions = ArticulationAction(joint_positions=target_joint_positions)
+            target_joint_positions = self._gripper_controller.forward(
+                action="open", current_joint_positions=current_joint_positions
+            )
         else:
-            target_joint_positions = super().forward(
-                current_joint_positions=current_joint_positions,
+            target_joint_positions = self._ik_solver.forward(
                 target_end_effector_position=position_target,
                 target_end_effector_orientation=euler_angles_to_quat(np.array([0, np.pi, 0])),
             )
@@ -127,10 +151,22 @@ class RMPFlowPickPlace(RMPFlowIKSolver):
     def _combine_convex(self, a, b, alpha):
         return (1 - alpha) * a + alpha * b
 
-    def reset(self):
-        super().reset()
+    def reset(self, start_picking_height=0.3, event_velocities=None):
+        BaseController.reset(self)
+        self._gripper_controller.reset()
+        self._ik_solver.reset()
         self._event = 0
         self._t = 0
+        self._h1 = start_picking_height
+        self._pause = False
+        if event_velocities is not None:
+            self._event_velocities = event_velocities
+            if not isinstance(self._event_velocities, np.ndarray) or not isinstance(self._event_velocities, list):
+                raise Exception("event velocities need to be list or numpy array")
+            elif isinstance(self._event_velocities, np.ndarray):
+                self._event_velocities = self._event_velocities.tolist()
+            if len(self._event_velocities) == 9:
+                raise Exception("event velocities need have length of 9")
         return
 
     def is_done(self):
@@ -138,3 +174,9 @@ class RMPFlowPickPlace(RMPFlowIKSolver):
             return True
         else:
             return False
+
+    def pause(self):
+        self._pause = True
+
+    def resume(self):
+        self._pause = False