+++ source/extensions/omni.isaac.samples/python/scripts/jetbot_keyboard/jetbot_keyboard.py
@@ -16,7 +16,7 @@ from omni.isaac.jetbot import Jetbot
 from omni.isaac.jetbot.controllers import DifferentialController
 from omni.isaac.core.utils.nucleus_utils import find_nucleus_server
 from omni.isaac.core.utils.stage import add_reference_to_stage
-from omni.isaac.core.tasks.task import BaseTask
+from omni.isaac.core.tasks import BaseTask
 from omni.isaac.core.scenes.scene import Scene
 from omni.isaac.samples.scripts.base_sample import BaseSample