+++ source/extensions/omni.isaac.ur10/omni/isaac/ur10/ur10.py
@@ -124,17 +124,17 @@ class UR10(Robot):
         self._gripper_length = length
         return
 
-    def _initialize_handles(self) -> None:
+    def initialize_handles(self) -> None:
         """[summary]
         """
-        super()._initialize_handles()
+        super().initialize_handles()
         self._end_effector_handle = self._dc_interface.find_articulation_body(
             self._handle, self._end_effector_prim_name
         )
         end_effector_prim_path = self._dc_interface.get_rigid_body_path(self._end_effector_handle)
         end_effector_prim = self._stage.GetPrimAtPath(end_effector_prim_path)
         self._end_effector = RigidPrim(prim=end_effector_prim, name=self._name + "_end_effector")
-        self._end_effector._initialize_handles()
+        self._end_effector.initialize_handles()
         # TODO: ask about gravity
         self.disable_gravity()
         return