+++ source/extensions/omni.isaac.core/docs/CHANGELOG.md
@@ -2,6 +2,21 @@
 CHANGELOG
 **********
 
+[0.1.2] - 2021-10-20
+========================
+
+Added
+-------
+- Added articulation gripper class
+- Deleted PD in the namings under articulation controller
+- Added base tasks for stacking, pick_place and follow_target
+- Added an IK solver under utils
+
+Changed
+-------
+- The behavior of .play() method under SimulationContext to always do 1 step with rendering for dc to function properly.
+
+
 [0.1.1] - 2021-10-18
 ========================